tcp: split TcpTxItem in its own header/implementation file
