+++ src/internet/wscript
@@ -162,6 +162,7 @@ def build(bld):
         'model/tcp-lp.cc',
         'model/tcp-rx-buffer.cc',
         'model/tcp-tx-buffer.cc',
+        'model/tcp-tx-item.cc',
         'model/tcp-option.cc',
         'model/tcp-option-rfc793.cc',
         'model/tcp-option-winscale.cc',
@@ -397,6 +398,7 @@ def build(bld):
         'model/tcp-socket-base.h',
         'model/tcp-socket-state.h',
         'model/tcp-tx-buffer.h',
+        'model/tcp-tx-item.h',
         'model/tcp-rx-buffer.h',
         'model/tcp-recovery-ops.h',
         'model/tcp-prr-recovery.h',
